src/main/java/org/example/algorithm/list/LinkedListCycleII.java
package org.example.algorithm.list;

import org.example.datastructure.list.ListNode;

public class LinkedListCycleII {

    public ListNode detectCycle(ListNode head) {
        if (head == null || head.next == null) {
            return null;
        }

        ListNode slow = head, fast = head;
        while (fast != null && fast.next != null) {
            fast = fast.next.next;
            slow = slow.next;
            if (slow == fast) {
                fast = head;
                break;
            }
        }

        if (fast != head) {
            return null;
        }

        while (fast != slow) {
            fast = fast.next;
            slow = slow.next;
        }

        return fast;
    }
}
